common release pasting/moving now working properly implemented support of multiCD case removed unnecessary classes

+++ src/Gears/FinalProcess.java
@@ -9,9 +9,9 @@ import java.io.File;
 import java.io.IOException;
 import static java.lang.Integer.valueOf;
 import static java.lang.System.out;
-import java.util.Map;
-import java.util.Set;
-import java.util.logging.Logger;
+import java.nio.file.DirectoryNotEmptyException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import org.apache.commons.io.FileUtils;
@@ -23,6 +23,7 @@ import org.jaudiotagger.audio.exceptions.InvalidAudioFrameException;
 import org.jaudiotagger.audio.exceptions.ReadOnlyFileException;
 import org.jaudiotagger.audio.mp3.MP3File;
 import org.jaudiotagger.tag.FieldDataInvalidException;
+import org.jaudiotagger.tag.FieldKey;
 import static org.jaudiotagger.tag.FieldKey.ALBUM;
 import static org.jaudiotagger.tag.FieldKey.ARTIST;
 import static org.jaudiotagger.tag.FieldKey.ARTIST_SORT;
@@ -34,7 +35,6 @@ import static org.jaudiotagger.tag.FieldKey.GENRE;
 import static org.jaudiotagger.tag.FieldKey.RECORD_LABEL;
 import static org.jaudiotagger.tag.FieldKey.TITLE;
 import static org.jaudiotagger.tag.FieldKey.TRACK;
-import static org.jaudiotagger.tag.FieldKey.TRACK_TOTAL;
 import static org.jaudiotagger.tag.FieldKey.YEAR;
 import org.jaudiotagger.tag.KeyNotFoundException;
 import org.jaudiotagger.tag.TagException;
@@ -55,6 +55,8 @@ public class FinalProcess {
     private String outputAlbumYearValue = "";
     private String outputAlbumTypeValue = "";
 
+    private final File destination;
+
     /**
      *
      * @param rootItem
@@ -62,6 +64,7 @@ public class FinalProcess {
     public FinalProcess(ItemProperties rootItem) {
         this.rootItem = rootItem;
         artwork = new StandardArtwork();
+        destination = new File(Controller.pathTextArea.getText());
     }
 
     /**
@@ -82,6 +85,12 @@ public class FinalProcess {
                     fixTags(medium);
                 }
             } else {
+                if (!rootItem.getMediumList().isEmpty()) {
+                    for (Medium medium : rootItem.getMediumList()) {
+                        fixTags(medium);
+                    }
+                }
+
                 for (ItemProperties child : rootItem.getChildList()) {
                     for (Medium medium : child.getMediumList()) {
                         fixTags(medium);
@@ -93,19 +102,27 @@ public class FinalProcess {
             System.out.println("Error occured while fixing tags process");
         }
 
-        if (rootItem.getChildList().isEmpty()) {
-            moveFactory(rootItem);
-        } else {
-            for (ItemProperties child : rootItem.getChildList()) {
-                moveFactory(child);
+        if (destination.exists()) {
+            if (rootItem.getChildList().isEmpty()) {
+                moveFactory(rootItem);
+                removeFileAndParentsIfEmpty(rootItem.getParentDir().getValue().toPath());
+            } else {
+                for (ItemProperties child : rootItem.getChildList()) {
+                    moveFactory(child);
+                    removeFileAndParentsIfEmpty(child.getCurrentDir().getValue().toPath());
+                }
+                removeFileAndParentsIfEmpty(rootItem.getParentDir().getValue().toPath());
             }
+
+        } else {
+            System.out.println(destination.getAbsolutePath() + " not exists");
         }
 
-        out.println(outputArtistValue);
-        out.println(outputAlbumYearValue);
-        out.println(outputAlbumTypeValue);
-        out.println(outputArtistGenresValue);
-        out.println(outputRecordGenresValue);
+        /*out.println(outputArtistValue);
+         out.println(outputAlbumYearValue);
+         out.println(outputAlbumTypeValue);
+         out.println(outputArtistGenresValue);
+         out.println(outputRecordGenresValue);*/
     }
 
     private void fixTags(Medium medium) throws KeyNotFoundException,
@@ -134,20 +151,22 @@ public class FinalProcess {
             file.getTag().setField(TRACK, trackString);
 
             //Запись тега TRACK_TOTAL
-            file.getTag().setField(TRACK_TOTAL, Integer.toString(medium.getAudioList().size()));
-
+            //file.getTag().setField(TRACK_TOTAL, Integer.toString(medium.getAudioList().size()));
             //Запись тега DISC_NO
             if (file.getTag().getFirst(DISC_NO).isEmpty()) {
-                if (medium.getCdN() > 0) {
-                    if (audio.getCdN() == 0 && medium.getCdN() != 0) {
-                        file.getTag().setField(DISC_NO, Integer.toString(medium.getCdN()));
-                    }
+                if (audio.getCdN() == 0 && rootItem.getCdN() > 0) {
+                    file.getTag().setField(DISC_NO, Integer.toString(medium.getCdN()));
                 }
             }
 
             //Запись тега ARTIST
-            if (!outputArtistValue.isEmpty() && !file.getTag().getFirst(ARTIST).equals(outputArtistValue)) {
+            if (!outputArtistValue.isEmpty() && !file.getTag().getFirst(ARTIST).equals(outputArtistValue)
+                    && !outputArtistValue.equals("Various Artists")) {
                 file.getTag().setField(ARTIST, outputArtistValue);
+            }
+
+            if (outputArtistValue.equals("Various Artists")) {
+                file.getTag().setField(FieldKey.IS_COMPILATION, Integer.toString(1));
             }
 
             //Запись тега ARTIST_SORT
@@ -281,10 +300,10 @@ public class FinalProcess {
             return "VA";
         } else {
             StringBuilder artistString = new StringBuilder();
-            artistString.append(validateName(outputArtistValue));
-
-            if (artistString.toString().startsWith("The ")) {
-                artistString.append(artistString.toString().replaceFirst("The ", "")).append(", The");
+            if (outputArtistValue.startsWith("The ")) {
+                artistString.append(validateName(outputArtistValue).replaceFirst("The ", "")).append(", The");
+            } else {
+                artistString.append(validateName(outputArtistValue));
             }
 
             if (!outputCountryValue.isEmpty() || !outputArtistGenresValue.isEmpty()) {
@@ -309,29 +328,58 @@ public class FinalProcess {
     private String buildAlbumString() {
         StringBuilder albumString = new StringBuilder();
 
-        rootItem.getMediumList().stream().map((medium) -> {
-            Set<Map.Entry<String, String>> set = medium.getYearAlbum().entrySet();
-            return medium;
-        }).forEach((medium) -> {
-            if (rootItem.hasVaAttribute()) { //если VA
-                albumString.append("VA")
-                        .append(" - ")
-                        .append(medium.getAlbum())
-                        .append(" (")
-                        .append(medium.getYear())
-                        .append(")");
-            } else { //если обычный релиз
-                if (selectedIssue != null) { //если выбрано издание
-                    albumString.append(selectedIssue.getIssueYear())
-                            .append(" - ")
-                            .append(selectedIssue.getIssueTitle());
-                } else { //если ничего не выбрано
-                    albumString.append(medium.getYear())
-                            .append(" - ")
-                            .append(medium.getAlbum());
+        if (!rootItem.getChildList().isEmpty()) {
+            for (ItemProperties child : rootItem.getChildList()) {
+                for (Medium medium : child.getMediumList()) {
+                    if (rootItem.hasVaAttribute()) { //если VA
+                        albumString.append("VA")
+                                .append(" - ")
+                                .append(medium.getAlbum())
+                                .append(" (")
+                                .append(medium.getYear())
+                                .append(")");
+                    } else { //если обычный релиз
+                        if (selectedIssue != null) { //если выбрано издание
+                            albumString.append(selectedIssue.getIssueYear())
+                                    .append(" - ")
+                                    .append(selectedIssue.getIssueTitle());
+                        } else { //если ничего не выбрано
+                            albumString.append(medium.getYear())
+                                    .append(" - ")
+                                    .append(medium.getAlbum());
+                        }
+                    }
+                }
+                if (rootItem.getCdN() > 0) {
+                    break;
                 }
             }
-        });
+        } else {
+            for (Medium medium : rootItem.getMediumList()) {
+                if (rootItem.hasVaAttribute()) { //если VA
+                    albumString.append("VA")
+                            .append(" - ")
+                            .append(medium.getAlbum())
+                            .append(" (")
+                            .append(medium.getYear())
+                            .append(")");
+                } else { //если обычный релиз
+                    if (selectedIssue != null) { //если выбрано издание
+                        albumString.append(selectedIssue.getIssueYear())
+                                .append(" - ")
+                                .append(selectedIssue.getIssueTitle());
+                    } else { //если ничего не выбрано
+                        albumString.append(medium.getYear())
+                                .append(" - ")
+                                .append(medium.getAlbum());
+                    }
+                }
+            }
+        }
+
+        if (rootItem.getCdN() > 0) {
+            albumString.append(" [").append(rootItem.getCdN()).append("CD]");
+        }
 
         return albumString.toString();
     }
@@ -343,46 +391,53 @@ public class FinalProcess {
             ReadOnlyFileException,
             InvalidAudioFrameException,
             CannotReadException {
-        
-        for (Medium medium : item.getMediumList()) {
-            int cdN = 0;
-            if (medium.getCdN() != 0) {
-                cdN = medium.getCdN();
-            } else if (rootItem.getCdN() != 0) {
-                cdN = rootItem.getCdN();
-            }
 
-            if (rootItem.hasVaAttribute()) {
-                File newAlbumDirectory = new File(Controller.pathTextArea.getText() + "\\" + buildAlbumString());
-                newAlbumDirectory.mkdir();
+        int cdN = 0;
 
-                if (cdN > 0) {
-                    File cdDirectory = new File(newAlbumDirectory + "\\" + "CD" + cdN);
-                    cdDirectory.mkdir();
-                    
-                    
-                    
-                    organiseAll(rootItem, cdDirectory);
-                } else {
-                    
-                }
+        if (item.getMediumList().size() == 1) {
+            cdN = findCdN(item.getMediumList().get(0));
+        } else if (item.getMediumList().size() > 1) {
+            System.out.println("Unhandeled case: getMediumList().size() > 1");
+        }
 
+        if (rootItem.hasVaAttribute()) {
+            File newAlbumDirectory = new File(destination + "\\" + buildAlbumString());
+            newAlbumDirectory.mkdir();
+
+            if (cdN > 0) {
+                File cdDirectory = new File(newAlbumDirectory + "\\" + "CD" + cdN);
+                cdDirectory.mkdir();
+                organiseAll(item, cdDirectory);
             } else {
-                File newArtistDirectory = new File(Controller.pathTextArea.getText() + "\\" + buildArtistString());
-                File newAlbumDirectory = new File(newArtistDirectory + "\\" + buildAlbumString());
-
-                newArtistDirectory.mkdir();
-                newAlbumDirectory.mkdir();
-
-                if (cdN > 0) {
-                    File cdDirectory = new File(newAlbumDirectory + "\\" + "CD" + cdN);
-                    cdDirectory.mkdir();
-
-                } else {
-
-                }
+                organiseAll(item, newAlbumDirectory);
             }
 
+            if (!rootItem.getChildList().isEmpty()) {
+                item = rootItem;
+            }
+
+            organiseAll(item, newAlbumDirectory);
+
+        } else {
+            File newArtistDirectory = new File(destination + "\\" + buildArtistString());
+            File newAlbumDirectory = new File(newArtistDirectory + "\\" + buildAlbumString());
+
+            newArtistDirectory.mkdir();
+            newAlbumDirectory.mkdir();
+
+            if (cdN > 0) {
+                File cdDirectory = new File(newAlbumDirectory + "\\" + "CD" + cdN);
+                cdDirectory.mkdir();
+                organiseAll(item, cdDirectory);
+            } else {
+                organiseAll(item, newAlbumDirectory);
+            }
+
+            if (!rootItem.getChildList().isEmpty()) {
+                item = rootItem;
+            }
+
+            organiseAll(item, newAlbumDirectory);
         }
 
     }
@@ -394,7 +449,7 @@ public class FinalProcess {
     }
 
     private void organiseAudio(ItemProperties ip, File to) {
-        if (ip.getListOfAudioFiles().isEmpty()) {
+        if (!ip.getListOfAudioFiles().isEmpty()) {
             for (AudioProperties audio : ip.getListOfAudioFiles()) {
                 move(audio.getFile(), to);
             }
@@ -402,7 +457,7 @@ public class FinalProcess {
     }
 
     private void organiseImages(ItemProperties ip, File to) {
-        if (ip.getListOfImageFiles().isEmpty()) {
+        if (!ip.getListOfImageFiles().isEmpty()) {
             if (ip.getListOfImageFiles().size() > 1) {
                 File coversFolder = new File(to + "\\" + "Covers");
                 coversFolder.mkdir();
@@ -423,11 +478,11 @@ public class FinalProcess {
     }
 
     private void organiseOthers(ItemProperties ip, File to) {
-        if (ip.getListOfOtherFiles().isEmpty()) {
+        if (!ip.getListOfOtherFiles().isEmpty()) {
             removeJunkFiles(ip.getListOfOtherFiles());
         }
 
-        if (ip.getListOfOtherFiles().isEmpty()) {
+        if (!ip.getListOfOtherFiles().isEmpty()) {
             for (File otherFile : ip.getListOfOtherFiles()) {
                 move(otherFile, to);
             }
@@ -450,14 +505,27 @@ public class FinalProcess {
 
     private void move(File from, File to) {
         try {
-            if (from.isDirectory()) {
-                FileUtils.moveDirectoryToDirectory(from, to, false);
-            } else {
-                FileUtils.moveFileToDirectory(from, to, false);
+            if (!to.exists()) {
+                if (from.isDirectory()) {
+                    FileUtils.moveDirectoryToDirectory(from, to, false);
+                } else {
+                    FileUtils.moveFileToDirectory(from, to, false);
+                }
             }
         } catch (IOException e) {
             System.out.println("move() exception\nFrom " + from + "\nTo " + to);
         }
+    }
+
+    private int findCdN(Medium medium) {
+        int cdN = 0;
+        if (medium.getCdN() != 0) {
+            cdN = medium.getCdN();
+        } else if (rootItem.getCdN() != 0) {
+            cdN = rootItem.getCdN();
+        }
+
+        return cdN;
     }
 
     //Определение переменных для записи в теги
@@ -498,6 +566,23 @@ public class FinalProcess {
         }
     }
 
+    public void removeFileAndParentsIfEmpty(Path path)
+            throws IOException {
+        if (path == null || path.endsWith(rootItem.getParentDir().getValue().getParentFile().getPath())) {
+            return;
+        }
+
+        if (Files.isDirectory(path)) {
+            try {
+                Files.delete(path);
+            } catch (DirectoryNotEmptyException e) {
+                return;
+            }
+        }
+
+        removeFileAndParentsIfEmpty(path.getParent());
+    }
+
     //Проверка существования обложки внутри аудиофайла
     private boolean audioArtworkExists(MP3File file) {
         return !file.getTag().getArtworkList().isEmpty();
@@ -519,7 +604,7 @@ public class FinalProcess {
             }
         } else {
             if (medium.getCurrentItem().getCdN() > 0) {
-
+                System.out.println("Unhandeled case: ");
             }
         }
         return exists;
@@ -563,6 +648,74 @@ public class FinalProcess {
     public void setRymp(RYMParser rymp) {
         this.rymp = rymp;
     }
-    private static final Logger LOG = Logger.getLogger(FinalProcess.class.getName());
 
+    /*public static <K extends Comparable<? super K>, V extends Comparable<? super V>> Map<K, V>
+     sortByValueThenByKey(Map<K, V> map) {
+     List<Map.Entry<K, V>> list
+     = new LinkedList<>(map.entrySet());
+     Collections.sort(list, new Comparator<Map.Entry<K, V>>() {
+     @Override
+     public int compare(Map.Entry<K, V> o1, Map.Entry<K, V> o2) {
+     return (o1.getValue()).compareTo(o2.getValue());
+     }
+     });
+
+     Collections.sort(list, new Comparator<Map.Entry<K, V>>() {
+     @Override
+     public int compare(Map.Entry<K, V> o1, Map.Entry<K, V> o2) {
+     return (o1.getKey()).compareTo(o2.getKey());
+     }
+     });
+
+     Map<K, V> result = new LinkedHashMap<>();
+     for (Map.Entry<K, V> entry : list) {
+     result.put(entry.getKey(), entry.getValue());
+     }
+     return result;
+     }/*
+
+     /*public boolean setCustomTag(AudioFile audioFile, String description, String text) {
+     FrameBodyTXXX txxxBody = new FrameBodyTXXX();
+     txxxBody.setDescription(description);
+     txxxBody.setText(text);
+
+     // Get the tag from the audio file
+     // If there is no ID3Tag create an ID3v2.3 tag
+     Tag tag = audioFile.getTagOrCreateAndSetDefault();
+     // If there is only a ID3v1 tag, copy data into new ID3v2.3 tag
+     if (!(tag instanceof ID3v23Tag || tag instanceof ID3v24Tag)) {
+     Tag newTagV23 = null;
+     if (tag instanceof ID3v1Tag) {
+     newTagV23 = new ID3v23Tag((ID3v1Tag) audioFile.getTag()); // Copy old tag data               
+     }
+     if (tag instanceof ID3v22Tag) {
+     newTagV23 = new ID3v23Tag((ID3v11Tag) audioFile.getTag()); // Copy old tag data              
+     }
+     audioFile.setTag(newTagV23);
+     }
+
+     AbstractID3v2Frame frame = null;
+     if (tag instanceof ID3v23Tag) {
+     frame = new ID3v23Frame("TXXX");
+     } else if (tag instanceof ID3v24Tag) {
+     frame = new ID3v24Frame("TXXX");
+     }
+
+     frame.setBody(txxxBody);
+
+     try {
+     tag.addField(frame);
+     } catch (FieldDataInvalidException e) {
+     e.printStackTrace();
+     return false;
+     }
+
+     try {
+     audioFile.commit();
+     } catch (CannotWriteException e) {
+     e.printStackTrace();
+     return false;
+     }
+     return true;
+     }*/
 }
